Major file realocation
